+++ Etyllica/src/br/com/etyllica/effects/GlobalEffect.java
@@ -14,7 +14,7 @@ public abstract class GlobalEffect extends Layer {
 
 	protected AnimationScript script;
 	
-	public GlobalEffect(float x, float y, float w, float h){
+	public GlobalEffect(int x, int y, int w, int h){
 		super(x,y,w,h);
 	}
