+++ Etyllica/src/br/com/etyllica/gui/button/ImageRoundButton.java
@@ -20,7 +20,7 @@ public class ImageRoundButton extends ImageButton{
 	}
 
 	@Override
-	public boolean onMouse(float mx, float my) {
+	public boolean onMouse(int mx, int my) {
 
 		return colideCircleCircle(mx-w/2, my-w/2, 1, 1);
 	}
